Tests: Distance, angle and dihedral angle in vectors
